Use enumerate instead of range(len) in split()
Enumerate spits out two objects, the count (or page number) and
something else I'm not sure of but is 100% not neeed so the variable
unused is literally unused when iterating.

One example of what unused returns:

(0, {'/Type': '/Page', '/Parent': IndirectObject(2, 0, 2114555952000), '/Resources': IndirectObject(6, 0, 2114555952000), '/Contents': [IndirectObject(5, 0, 2114555952000)], '/MediaBox': [0, 0, 792, 612], '/CropBox': [0, 0, 792, 612]})

But not sure what the type is?