Rewrite Prep Deluxe PO to use UploadView and DownloadView
Originally, the user would upload the .xlsx template file first and
then the .pdf invoices. I couldn't quite figure out how to use the same
view with different redirects so instead just made it so the user can
upload both file types at once for this function.

Since both file types are uploaded to the upload folder, I had to figure
out how to download just the required file (in this case, the .xlsx) and
decided to just create a new download folder and modify the prep_po
function to save the completed file in that directory.

While simple, this does break the previous functions so I need to fix
those!